+++ HISTORY.rst
@@ -6,6 +6,9 @@ History
 ------------------
 
 * Fix Comments notification on Slack (ericof).
+* Add order.wf.edit_requirements event (ericof).
+* Add CustomerUserProfile and BriefyUserProfile events (ericof).
+* Add Slack actions to CustomerUserProfile and BriefyUserProfile (ericof).
 
 2.0.1 (2017-02-14)
 ------------------